Decider: ideology shift
The original method of computing the expected value of hitting a given
hand was to hit and then recursively pull the expected value of the
resulting hand out of expectedHitMap.

With this idea, the value of hitting, for example, a hard 20 against a
dealer ace can be found by performing simulations and noting the
following:
if the player busts, he wins 0 units on that simulation; however, if he
survives the hit (by getting an ace), the resulting hand is a hard 21
and is assigned a win of ~0.922 units (the expected value of standing on
a hard 21 vs an ace).

The problem with this approach is that simulation errors propagate down
the recursion tree, leading to bad results on hands that are either soft
or have low value.

The solution is to just use these values to simulate best play. For
example, on a hard 21 vs an ace, simulations (and common sense) say to
stay. Hitting a hard 20 either results in busting or a hard 21, which
can be played out with the best strategy for a hard 21 (staying). This
idea then extends to all other hands.